chore: refactor controllers

+++ src/controllers/DeleteUserByIdController.js
@@ -1,13 +1,22 @@
 const User = require('../models/User')
 
 class DeleteUserByIdController {
-  handle(request, response){
+  async handle(request, response){
     try {
       const { userId } = request.params
+
+      const user = await User.findOne({
+        'login.uuid': userId
+      }).exec()
+
+      if(!user){
+        return response.status(404).json({ error: 'User not found' })
+      }
+
       User.findOneAndDelete({ 'login.uuid': userId })
       response.status(200).json({ message: 'User deleted successfully' })
     } catch (error){
-      return response.status(404).json({ error: 'User not found' })
+      return response.status(500).json({ error: 'Internal server error. Please, try again later' })
     }
   }
 }
